refactor: split connector action boundaries
